[cmake+fs] move fs-backend's test to test/fs
