+++ test/Runner.cpp
@@ -33,21 +33,21 @@
 #include "TestValidate.hpp"
 #include "TestReadOnlyHDF5.hpp"
 #include "TestGroupHDF5.hpp"
-#include "TestAttributesFS.hpp"
-#include "TestFileFS.hpp"
-#include "TestBlockFS.hpp"
-#include "TestEntityFS.hpp"
-#include "TestEntityWithMetadataFS.hpp"
-#include "TestEntityWithSourcesFS.hpp"
-#include "TestFeatureFS.hpp"
-#include "TestGroupFS.hpp"
-#include "TestMultiTagFS.hpp"
-#include "TestPropertyFS.hpp"
-#include "TestDataArrayFS.hpp"
-#include "TestSectionFS.hpp"
-#include "TestSourceFS.hpp"
-#include "TestTagFS.hpp"
-#include "TestBaseTagFS.hpp"
+#include "fs/TestAttributesFS.hpp"
+#include "fs/TestFileFS.hpp"
+#include "fs/TestBlockFS.hpp"
+#include "fs/TestEntityFS.hpp"
+#include "fs/TestEntityWithMetadataFS.hpp"
+#include "fs/TestEntityWithSourcesFS.hpp"
+#include "fs/TestFeatureFS.hpp"
+#include "fs/TestGroupFS.hpp"
+#include "fs/TestMultiTagFS.hpp"
+#include "fs/TestPropertyFS.hpp"
+#include "fs/TestDataArrayFS.hpp"
+#include "fs/TestSectionFS.hpp"
+#include "fs/TestSourceFS.hpp"
+#include "fs/TestTagFS.hpp"
+#include "fs/TestBaseTagFS.hpp"
 
 
 int main(int argc, char* argv[]) {
